% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sysreqs.R
\name{get_sysreqs}
\alias{get_sysreqs}
\title{Get system requirements by R package}
\usage{
get_sysreqs(packages, quiet = TRUE, batch_n = 30)
}
\arguments{
\item{packages}{character vector of packages names}

\item{quiet}{boolean if TRUE the function is quiet}

\item{batch_n}{number of simultaneous packages to ask}
}
\value{
character vector of sysreqs
}
\description{
Get system requirements by R package
}
