% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pkgdown.R
\name{use_open_pkgdown}
\alias{use_open_pkgdown}
\title{Create an \code{open_pkgdown} function}
\usage{
use_open_pkgdown(
  pkgname = NULL,
  save_as = "R/open_pkgdown.R",
  Rcreate = FALSE,
  ...
)
}
\arguments{
\item{pkgname}{package name}

\item{...}{
  Arguments passed on to \code{\link[usethis:use_template]{usethis::use_template}}
  \describe{
    \item{\code{template}}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}
    \item{\code{save_as}}{Path of file to create, relative to root of active project.
Defaults to \code{template}}
    \item{\code{data}}{A list of data passed to the template.}
    \item{\code{ignore}}{Should the newly created file be added to \code{.Rbuildignore}?}
    \item{\code{open}}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
    \item{\code{package}}{Name of the package where the template is found.}
  }}

\item{create}{logical - should pkgdown be created?}
}
\value{
invisible
}
\description{
Create an \code{open_pkgdown} function
}
