% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimstemplates-package.R, R/package.R
\docType{package}
\name{jimstemplates-package}
\alias{jimstemplates}
\alias{jimstemplates-package}
\title{jimstemplates: Collection of templates}
\description{
What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimbrig2011@outlook.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})


\strong{Maintainer}: Jimmy Briggs \email{jimbrig2011@outlook.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
